+++ apps/client/src/app/pages/features/features-page.module.ts
@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
+import { GfPremiumIndicatorModule } from '@ghostfolio/ui/premium-indicator';
 
 import { FeaturesPageRoutingModule } from './features-page-routing.module';
 import { FeaturesPageComponent } from './features-page.component';
@@ -9,8 +10,9 @@ import { FeaturesPageComponent } from './features-page.component';
 @NgModule({
   declarations: [FeaturesPageComponent],
   imports: [
-    FeaturesPageRoutingModule,
     CommonModule,
+    FeaturesPageRoutingModule,
+    GfPremiumIndicatorModule,
     MatButtonModule,
     MatCardModule
   ],